pyproject.toml
[project]
name = "zorkdemo"
version = "0.3.1"
description = "A (much simplified) port of a famous adventure game to help teach my daughter how to program in Python."
authors = [{ name = "Brandon Corfman", email = "bcorfman@fastmail.fm" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.0,<2",
    "environs>=9.5.0,<15",
    "fastapi>=0.115.0,<1",
    "Markdown>=3.4.1,<4",
    "psycopg[binary]>=3.2.0,<4",
    "rich>=13.0.0,<15",
    "SQLAlchemy>=2.0.35,<3",
    "uvicorn>=0.32.0,<1",
    "yazm-py",
]

[dependency-groups]
dev = [
    "coverage>=7.0.1,<8",
    "httpx>=0.28.0,<1",
    "ipython>=8.7.0,<9",
    "pyinstaller>=6.0.0,<7",
    "pytest>=7.2.0,<10",
    "ruff>=0.11.0,<1",
]

[tool.setuptools]
py-modules = ["zorkdemo"]

[tool.setuptools.packages.find]
where = ["."]
include = ["adventure*", "backend*"]

[tool.uv]
default-groups = ["dev"]

[tool.uv.sources]
yazm-py = { path = "../yazm-py", editable = true }

[tool.ruff]
line-length = 127
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests", "backend/tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
